+++ src/myoddweb.directorywatcher.test/WatcherAddTests.cs
@@ -58,5 +58,76 @@ namespace myoddweb.directorywatcher.test
       }
       Assert.AreEqual(numberToAdd, numberAdded);
     }
+
+    [TestCase(true)]
+    [TestCase(false)]
+    public async Task StartWatcherBeforeAddingAnything( bool recursive)
+    {
+      using var helper = new HelperTest();
+      using var watcher = new Watcher();
+
+      // start the watcher
+      watcher.Start();
+
+      // the number of files we actually added
+      var numberAdded = 0;
+      const int numberToAdd = 10;
+      var timeout = (numberToAdd <= 2 ? 3 : numberToAdd) * 1000;
+      Assert.IsTrue(SpinWait.SpinUntil(() => watcher.Ready(), timeout));
+
+      // and then add the folder we are after
+      watcher.Add(new Request(helper.Folder, recursive));
+
+      // then wait
+      Assert.IsTrue(SpinWait.SpinUntil(() => watcher.Ready(), timeout));
+
+      watcher.OnAddedAsync += (fse, token) =>
+      {
+        Interlocked.Increment(ref numberAdded);
+        return Task.CompletedTask;
+      };
+
+      TestContext.Out.WriteLine("All watchers ready!");
+
+      for (var i = 0; i < numberToAdd; ++i)
+      {
+        // create an empty file
+        helper.AddFile();
+        await Task.Yield();
+      }
+
+      // give a bit of time
+      SpinWait.SpinUntil(() => numberAdded >= numberToAdd, timeout);
+    }
+
+    [Test]
+    public void IfWeHaveNoFoldersToWatchThenStartReturnsFalse()
+    {
+      using var helper = new HelperTest();
+      using var watcher = new Watcher();
+
+      // start the watcher, we return false as we have not folders.
+      Assert.IsFalse(watcher.Start());
+
+      // stop it.
+      watcher.Stop();
+    }
+
+    [Test]
+    public void StartedWatcherWithNoFolderIsAlwaysReady()
+    {
+      using var helper = new HelperTest();
+      using var watcher = new Watcher();
+
+      // start the watcher
+      watcher.Start();
+
+      // nothing was added, but we are ready
+      Assert.IsTrue(watcher.Ready());
+
+      // stop it.
+      watcher.Stop();
+    }
   }
+
 }